adding the jwt authentication

+++ controllers/api/v1/posts_api.js
@@ -25,29 +25,31 @@ module.exports.index = async function(req,res){
 module.exports.destroy = async function (req, res) {
     try {
         let post = await Post.findById(req.params.id);
-        // if (post.user == req.user.id) {
+        if (post.user == req.user.id) {
             post.remove();
             await Comment.deleteMany(
                 { post: req.params.id });
-            // if (req.xhr) {
-            //     return res.status(200).json({
-            //         data: {
-            //             post_id: req.params.id
-            //         },
-            //         message: "Post deleted"
-            //     });
-            // }
+            if (req.xhr) {
+                return res.status(200).json({
+                    data: {
+                        post_id: req.params.id
+                    },
+                    message: "Post deleted"
+                });
+            }
 
             // req.flash('success', 'Post and associated comments deleted!');
 
             return res.json(5000,{
                 message : "Internal Server Error"
             });
-        // }
-        // else {
-            // req.flash('error', err);
-            // return res.redirect('back');
-        // }
+        }
+        else {
+            req.flash('error', err);
+            return res.json(401,{
+                message : "You can not delete this post"
+            });
+        }
     } catch (err) {
         console.log(err);
         return;
